fix(authenticate): made a phone authentication "strategy"
